backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
nltk==3.8.1
scikit-learn==1.3.2
sentence-transformers==2.2.2
pdfminer.six==20221105
python-docx==1.0.1
aiofiles==23.2.1
numpy==1.26.2
